% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENSEXONS_37}
\alias{ENSEXONS_37}
\title{Ensembl exons build HG37 (GRCh37) v40 from gencode}
\format{
A data frame  with 62,739 rows and 7 variables:
\describe{
\item{chrom}{Chromosome on build version 38 (GRCh38/hg38)}
\item{gene_start}{genetic position of gene start on build version 38}
\item{gene_end}{genetic position of gene end on build version 38 }
\item{gene_symbol}{The name of the gene}
\item{exon_chromstart}{genetic positions of exon start}
\item{exon_chromend}{genetic position of exon end}
}
}
\usage{
ENSEXONS_37
}
\description{
exons on chrY and chrM were excluded from the exon dataset
}
\keyword{datasets}
