% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENSGENES_37}
\alias{ENSGENES_37}
\title{Ensembl genes build HG37 (GRCh37) v40 from gencode}
\format{
A data frame  with 62,743 rows and 5 variables:
\describe{
\item{chrom}{Chromosome on build version 38 (GRCh38/hg38)}
\item{gene_start}{genetic position of gene start on build version 38}
\item{gene_end}{genetic position of gene end on build version 38 }
\item{gene_symbol}{The name of the gene}
\item{biotype}{the biotype of the gene}
}
}
\usage{
ENSGENES_37
}
\description{
genes on chrY and chrM were excluded
}
\keyword{datasets}
