% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toprdata.R
\docType{package}
\name{toprdata}
\alias{toprdata}
\title{toprdata}
\description{
Gene and exon datasets to use with the topr package
}
\section{datasets}{

The two datasets are:
\itemize{
\item \code{\link{ENSGENES}} Gene information
\item \code{\link{ENSEXONS}} Exon informatino
}
}

\examples{
library(toprdata)
head(ENSGENES)
head(ENSEXONS)
}
